Replace reset_pcl_msg with point_cloud_msg_wrapper

+++ src/drivers/velodyne_nodes/include/velodyne_nodes/velodyne_cloud_node.hpp
@@ -87,7 +87,7 @@ private:
   uint32_t m_point_cloud_idx;
   autoware::common::lidar_utils::PointCloudIts m_point_cloud_its;
   const std::string m_frame_id;
-  const std::size_t m_cloud_size;
+  const std::uint32_t m_cloud_size;
 };  // class VelodyneCloudNode
 
 using VLP16DriverNode = VelodyneCloudNode<velodyne_driver::VLP16Data>;
